Finished Questions 2-4 and Finalize Analysisd agruments
To being with, I asked if there was a positive correlation between average costs and ratings, and there was no evidence to support nor deny that hypothesis. Then I looked to see what restuarant was the best based on ratings with a vote ratio being the tiebreaker. I was able to determine that ID 4728 is the best restuarant. In question 3, I looked for the lowest priced and quickest cooked place. Based on the table, we can determine that restuarant 1368, has the lowest average pricing and has the fastest cook time. Finally, I looked for a correlation between ratings and cook times, and in like analysis 1, there was no evidence supporting nor denying that there is a correlation.